Added swerve aiming idea, removed useless members, updated vendordeps
This aiming mode is designed specifically for scoring into the trough. It is explained in the comments.
I added some constants and a helper method to facilitate the aiming.
The useless members I removed were the rotational deadbands in the custom swerve requests. I did this because we don't need to worry about controller drift when calculating the rotational rate. A deadband would only interfere with the motion profile and controller.
I also fixed a bug where we added vision measurements in the wrong time base. See this link for more information:
https://www.chiefdelphi.com/t/ctre-swerve-addvisionmeasurment-having-no-effect/482024/5