Fixed motion profile reset and driver orientation reset, updated code formatter
The previous implementation of motion profile resetting checked whether to reset the profile during robot startup, which is very incorrect. I added a separate method for profiled requests that resets the profile without having to check for a motion profile.
For some reason, CTRE changed the seedFieldCentric method so it now resets the robot rotation in addition to the driver orientation. This causes the robot to lie about its current rotation (which we don't want) so I changed the button to only reset the driver orientation.