Renamed paths to use labeled tree names...
and continued working on AutoCreator.
I also removed the constraint zones from the paths because we will need to limit the acceleration whenever the elevator is up, which is all the time in autonomous. If we get any inaccuracies, we can incorporate our custom swerve request into AutoCreator like with our teleop auto-alignment.